move shared http code into utils, remove request from tables node
